Travis tweaks; 5.5.0 operators special handling
Avoid building the bundle twice, by not running it automatically via NPM prepare.

Add a small workaround to produce a System bundle that includes the ability to:

import { map } from 'rxjs/operators';

This workaround includes all of the operators in the system bundle, which is ideals for our purpose here. The documentation includes a much more complex process for bundling just the desired operators using a Webpack build.